+++ cloudify/test_utils/mock_rest_client.py
