--- cloudify/tests/mocks/mock_module.py
@@ -1,20 +1,0 @@
-########
-# Copyright (c) 2016 GigaSpaces Technologies Ltd. All rights reserved
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
-def mock_hook_function(context, **kwargs):
-    with open('/tmp/mock_hook_function.txt', 'a') as f:
-        f.write("In mock_hook_function, context: {0} kwargs: {1}\n"
-                .format(context, kwargs))
